project/pyproject.toml
[project]
name = "instagrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=5.2,<5.3",
    "django-cors-headers>=4.9.0",
    "django-stubs-ext>=5.2.9",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "faker>=40.1.2",
    "markitdown>=0.1.4",
]

[dependency-groups]
dev = [
    "celery-types>=0.24.0",
    "django-stubs[compatible-mypy]>=5.2.9",
    "djangorestframework-stubs[compatible-mypy]>=3.16.8",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "instagrader.settings"
python_files = ["tests.py", "test_*.py"]
